// src/constants/common.interface.ts
export enum UserRole {
	Student = 'student',
	Admin = 'admin',
}

export enum Gender {
	Male = 'male',
	Female = 'female',
	Other = 'other',
}

export enum PaidIn {
	Cash = 'cash',
	Card = 'card',
	Cheque = 'cheque',
	Online = 'online',
	UPI = 'upi',
	phonePay = 'phonePay',
	GooglePay = 'googlePay',
	Paytm = 'paytm',
	Others = 'others',
}

export enum ExpenseCategory {
	FOOD = 'Food',
	TRANSPORT = 'Transport',
	ENTERTAINMENT = 'Entertainment',
	HEALTH = 'Health',
	UTILITIES = 'Utilities',
	HOUSING = 'Housing',
	GROCERIES = 'Groceries',
	INSURANCE = 'Insurance',
	EDUCATION = 'Education',
	SAVINGS = 'Savings',
	INVESTMENTS = 'Investments',
	TRAVEL = 'Travel',
	GIFTS = 'Gifts',
	DONATIONS = 'Donations',
	SUBSCRIPTIONS = 'Subscriptions',
	CLOTHING = 'Clothing',
	PERSONAL_CARE = 'Personal Care',
	RESTAURANTS = 'Restaurants',
	PETS = 'Pets',
	CHILDCARE = 'Childcare',
	MISCELLANEOUS = 'Miscellaneous',
	DEBT = 'Debt',
	EMERGENCY = 'Emergency',
	FUN = 'Fun',
	INTERNET = 'Internet',
	PHONE = 'Phone',
	WATER = 'Water',
	ELECTRICITY = 'Electricity',
	GAS = 'Gas',
	MAINTENANCE = 'Maintenance',
	PROPERTY_TAX = 'Property Tax',
	RENT = 'Rent',
	MORTGAGE = 'Mortgage',
	MEDICAL = 'Medical',
	FITNESS = 'Fitness',
	BEAUTY = 'Beauty',
	ENTERTAINMENT_SUBSCRIPTIONS = 'Entertainment Subscriptions',
	HOBBIES = 'Hobbies',
	PROFESSIONAL_FEES = 'Professional Fees',
	BUSINESS_EXPENSES = 'Business Expenses',
	HOME_IMPROVEMENT = 'Home Improvement',
	OFFICE_SUPPLIES = 'Office Supplies',
	TAXES = 'Taxes',
	LEGAL_FEES = 'Legal Fees',
	CHARITY = 'Charity',
	RETIREMENT = 'Retirement',
	VACATION = 'Vacation',
	FESTIVALS = 'Festivals',
}